laatste wijzigingen aan infra rol

+++ roles/infra.sh
@@ -9,7 +9,6 @@ function dhcpserver {
         cp ./roles/files/system/dhcpd${dhcpProtocol}.conf /etc/dhcp/
     fi
 
-    echo "Setup DHCP service Units"
     if [[ ! -f /etc/systemd/system/dhcp-server${dhcpProtocol}.service ]]; then
         echo "DHCP IPv$dhcpProtocol Unit"
         cp ./roles/files/system/dhcp-server${dhcpProtocol}.service /etc/systemd/system/
@@ -32,7 +31,9 @@ apt update
 apt install -y bind9 bind9-dnsutils bind9-doc bind9-host isc-dhcp-server curl wget gpg apt-transport-https 
 
 echo "Toevoegen van elasticsearch key aan repository"
-wget -qO - https://artifacts.elastic.co/GPG-KEY-elasticsearch | sudo gpg --dearmor -o /usr/share/keyrings/elasticsearch-keyring.gpg
+if [[ ! -f /usr/share/keyrings/elasticsearch-keyring.gpg ]]; then
+    wget -qO - https://artifacts.elastic.co/GPG-KEY-elasticsearch | sudo gpg --dearmor -o /usr/share/keyrings/elasticsearch-keyring.gpg
+fi
 
 echo "Bijwerken sources list en bijwerken cache"
 echo "deb [signed-by=/usr/share/keyrings/elasticsearch-keyring.gpg] https://artifacts.elastic.co/packages/8.x/apt stable main" | sudo tee /etc/apt/sources.list.d/elastic-8.x.list
